+++ changelog.md
@@ -20,6 +20,7 @@
 * Added shape property to `Scan` and `Kymo`.
 * Added a warning to the Kymotracker widget if the threshold parameter is set too low, which may result in slow tracking and the widget hanging.
 * Pylake now depends on `h5py>=3.4, <4`. This change is required to still support `len()` with the lazy loading fix for `TimeSeries`.
+* Added header line to exported track coordinate CSV files. The first header line now contains the version of `pylake` which generated the file and a version number for the CSV file itself (starting with `v2` from this release).
 
 #### Bug fixes